added working code
